+++ source/library.h
@@ -4,6 +4,8 @@
 #include <ctime>
 #include <cstdlib>
 #include <string.h>
+#include<vector>
+#include<vector>
 using namespace std;
 
 // I. Algorithms
@@ -93,7 +95,7 @@ void writingFile (string outputFile, int a[], int n);
 double getTimeOfAlg (char *alg, int a[], int n);
 
 // 4. get number of comparisons of the algorithm
-unsigned long int getCompOfAlg (char *alg, int a[], int n);
+unsigned long long getCompOfAlg (char *alg, int a[], int n);
 
 // 5. doOutputOrder function
 void doOutputOrder (char *output, double &time, unsigned long long &comp, int a[], int n, char *alg);
